Inject terminal runtime filesystem seams
