+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/FakeHibernationRecorder.swift
@@ -1,0 +1,6 @@
+import Foundation
+@testable import CmuxTerminal
+
+final class FakeHibernationRecorder: AgentHibernationRecording {
+    func recordTerminalInput(workspaceId: UUID, panelId: UUID) {}
+}
