+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/FakeRendererRealizationScheduler.swift
@@ -1,0 +1,6 @@
+@testable import CmuxTerminal
+
+final class FakeRendererRealizationScheduler: TerminalRendererRealizationScheduling {
+    @MainActor
+    func scheduleImmediatePass() {}
+}
