+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/FakeTerminalByteTeeLease.swift
@@ -1,0 +1,5 @@
+@testable import CmuxTerminal
+
+final class FakeTerminalByteTeeLease: TerminalByteTeeLease {
+    func release() {}
+}
