+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/FakeTerminalSurfaceViewProvider.swift
@@ -1,0 +1,15 @@
+import AppKit
+@testable import CmuxTerminal
+
+@MainActor
+struct FakeTerminalSurfaceViewProvider: TerminalSurfaceViewProviding {
+    let surfaceView: FakeTerminalSurfaceNativeView
+    let paneHost: FakeTerminalSurfacePaneHost
+
+    func makeSurfaceViews(
+        initialFrame: NSRect
+    ) -> (surfaceView: any TerminalSurfaceNativeViewing, paneHost: any TerminalSurfacePaneHosting) {
+        _ = initialFrame
+        return (surfaceView, paneHost)
+    }
+}
