+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/ManualRestoreSpawnDelay.swift
@@ -1,0 +1,6 @@
+@testable import CmuxTerminal
+
+@MainActor
+final class ManualRestoreSpawnDelay: TerminalSurfaceRestoreSpawnDelayCancelling {
+    func cancel() {}
+}
