+++ Packages/CmuxTerminal/Tests/CmuxTerminalTests/TerminalSurfaceRestoreSpawnSchedulerTests.swift
@@ -6,47 +6,6 @@ import CmuxTerminalCore
 @testable import CmuxTerminal
 
 @MainActor
-private final class ManualRestoreSpawnDelayer: TerminalSurfaceRestoreSpawnDelaying {
-    private var delayCount = 0
-    private var delayOperations: [@MainActor () -> Void] = []
-    private var delayOperationHead = 0
-    private var countContinuations: [Int: [CheckedContinuation<Void, Never>]] = [:]
-
-    func scheduleDelay(
-        for duration: Duration,
-        operation: @escaping @MainActor () -> Void
-    ) -> any TerminalSurfaceRestoreSpawnDelayCancelling {
-        _ = duration
-        delayCount += 1
-        let count = delayCount
-        for waiter in countContinuations.removeValue(forKey: count) ?? [] {
-            waiter.resume()
-        }
-        delayOperations.append(operation)
-        return ManualRestoreSpawnDelay()
-    }
-
-    func waitForDelayCount(_ count: Int) async {
-        guard delayCount < count else { return }
-        await withCheckedContinuation { continuation in
-            countContinuations[count, default: []].append(continuation)
-        }
-    }
-
-    func releaseNextDelay() {
-        guard delayOperationHead < delayOperations.count else { return }
-        let operation = delayOperations[delayOperationHead]
-        delayOperationHead += 1
-        operation()
-    }
-}
-
-@MainActor
-private final class ManualRestoreSpawnDelay: TerminalSurfaceRestoreSpawnDelayCancelling {
-    func cancel() {}
-}
-
-@MainActor
 @Suite struct TerminalSurfaceRestoreSpawnSchedulerTests {
     @Test func restoredSurfaceSpawnsDrainOnePerDelay() async {
         let delayer = ManualRestoreSpawnDelayer()
@@ -348,145 +307,15 @@ private final class ManualRestoreSpawnDelay: TerminalSurfaceRestoreSpawnDelayCan
                 hibernationRecorder: FakeHibernationRecorder(),
                 runtimeTeardown: TerminalSurfaceRuntimeTeardownCoordinator(),
                 restoreSpawnScheduler: scheduler,
+                runtimeFilesystem: TerminalSurfaceRuntimeFilesystem(
+                    claudeCommandShimTemporaryDirectory: URL(fileURLWithPath: "/tmp/cmux-terminal-tests", isDirectory: true),
+                    installClaudeCommandShim: { _, _, _ in nil },
+                    isExecutableFile: { _ in false }
+                ),
                 sessionPortBase: 40_000,
                 sessionPortRangeSize: 100,
                 scrollbackReplayEnvironmentKey: "CMUX_TEST_SCROLLBACK_REPLAY"
             )
         )
     }
-}
-
-@MainActor
-private final class RecordingRestoreSpawnScheduler: TerminalSurfaceRuntimeSpawnScheduling {
-    private(set) var scheduledSurfaceIds: [UUID] = []
-    private var scheduledOperations: [@MainActor () -> Void] = []
-
-    func scheduleRestoredSurfaceSpawn(surfaceId: UUID, operation: @escaping @MainActor () -> Void) {
-        scheduledSurfaceIds.append(surfaceId)
-        scheduledOperations.append(operation)
-    }
-
-    func runScheduledOperation(at index: Int = 0) {
-        scheduledOperations[index]()
-    }
-}
-
-private final class FakeSurfaceRegistry: TerminalSurfaceRegistering {
-    func register(_ surface: any TerminalSurfacing) {}
-    func unregister(_ surface: any TerminalSurfacing) {}
-    func registerRuntimeSurface(_ surface: ghostty_surface_t, ownerId: UUID) {}
-    func unregisterRuntimeSurface(_ surface: ghostty_surface_t, ownerId: UUID) {}
-    func runtimeSurfaceOwnerId(_ surface: ghostty_surface_t) -> UUID? { nil }
-    func surface(id: UUID) -> (any TerminalSurfacing)? { nil }
-    func isRightSidebarDockSurface(id: UUID) -> Bool { false }
-    func allSurfaces() -> [any TerminalSurfacing] { [] }
-}
-
-@MainActor
-private final class FakeTerminalEngine: TerminalEngineHosting {
-    var runtimeApp: ghostty_app_t? { nil }
-    var runtimeConfig: ghostty_config_t? { nil }
-    var userGhosttyShellIntegrationMode: String { "none" }
-}
-
-@MainActor
-private struct FakeTerminalSurfaceViewProvider: TerminalSurfaceViewProviding {
-    let surfaceView: FakeTerminalSurfaceNativeView
-    let paneHost: FakeTerminalSurfacePaneHost
-
-    func makeSurfaceViews(
-        initialFrame: NSRect
-    ) -> (surfaceView: any TerminalSurfaceNativeViewing, paneHost: any TerminalSurfacePaneHosting) {
-        _ = initialFrame
-        return (surfaceView, paneHost)
-    }
-}
-
-private final class FakeTerminalSurfaceNativeView: NSView {
-    var tabId: UUID?
-    var hostedTabId: UUID? { tabId }
-    weak var attachedController: (any TerminalSurfaceControlling)?
-    var attachedSurfaceController: (any TerminalSurfaceControlling)? { attachedController }
-    var currentKeyStateIndicatorText: String? { nil }
-    var isKeyboardCopyModeActive: Bool { false }
-
-    func toggleKeyboardCopyMode() -> Bool { false }
-    func applyWindowBackgroundIfActive() {}
-    func forceRefreshSurface() -> Bool { true }
-}
-
-extension FakeTerminalSurfaceNativeView: @MainActor TerminalSurfaceHosting {}
-extension FakeTerminalSurfaceNativeView: @MainActor TerminalSurfaceNativeViewing {}
-
-@MainActor
-private final class FakeTerminalSurfacePaneHost: NSView, TerminalSurfacePaneHosting {
-    private let surfaceView: FakeTerminalSurfaceNativeView
-
-    init(surfaceView: FakeTerminalSurfaceNativeView) {
-        self.surfaceView = surfaceView
-        super.init(frame: surfaceView.frame)
-        addSubview(surfaceView)
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) is unavailable in tests")
-    }
-
-    func attachSurface(_ surface: TerminalSurface) {
-        surfaceView.attachedController = surface
-    }
-
-    func cancelFocusRequest() {}
-    func setVisibleInUI(_ visible: Bool) {}
-    func setActive(_ active: Bool) {}
-    func syncKeyStateIndicator(text: String?) {}
-    func setMobileViewportBorder(size: CGSize?, drawRight: Bool, drawBottom: Bool) {}
-}
-
-@MainActor
-private final class FakeSpawnPolicyProvider: TerminalSurfaceSpawnPolicyProviding {
-    func currentSpawnPolicy() -> TerminalSurfaceSpawnPolicy {
-        TerminalSurfaceSpawnPolicy(
-            claudeHooksEnabled: true,
-            customClaudePath: nil,
-            subagentNotificationEnvironmentKey: "CMUX_TEST_SUPPRESS_SUBAGENT_NOTIFICATIONS",
-            suppressSubagentNotifications: false,
-            cursorHooksEnabled: true,
-            geminiHooksEnabled: true,
-            kiroHooksEnabled: true,
-            kiroNotificationLevel: "all",
-            ampHooksEnabled: true,
-            shellIntegrationEnabled: false,
-            watchGitStatusEnabled: false,
-            showPullRequestsEnabled: false
-        )
-    }
-
-    func controlSocketPath() -> String {
-        "/tmp/cmux-test.sock"
-    }
-}
-
-private final class FakeTerminalByteTee: TerminalByteTeeBinding {
-    @MainActor
-    func installTee(on surface: ghostty_surface_t, surfaceID: UUID) -> any TerminalByteTeeLease {
-        FakeTerminalByteTeeLease()
-    }
-
-    @MainActor
-    func dropSurface(surfaceID: UUID) {}
-}
-
-private final class FakeTerminalByteTeeLease: TerminalByteTeeLease {
-    func release() {}
-}
-
-private final class FakeRendererRealizationScheduler: TerminalRendererRealizationScheduling {
-    @MainActor
-    func scheduleImmediatePass() {}
-}
-
-private final class FakeHibernationRecorder: AgentHibernationRecording {
-    func recordTerminalInput(workspaceId: UUID, panelId: UUID) {}
 }
