+++ Sources/TerminalSurfaceRuntimeWiring.swift
@@ -127,6 +127,25 @@ final class TerminalAgentHibernationRecorder: AgentHibernationRecording {
     }
 }
 
+// MARK: Filesystem
+
+extension TerminalSurfaceRuntimeFilesystem {
+    static func live() -> TerminalSurfaceRuntimeFilesystem {
+        TerminalSurfaceRuntimeFilesystem(
+            claudeCommandShimTemporaryDirectory: FileManager.default.temporaryDirectory,
+            installClaudeCommandShim: {
+                TerminalSurface.installClaudeCommandShimIfPossible(
+                    wrapperURL: $0,
+                    surfaceId: $1,
+                    temporaryDirectory: $2,
+                    fileManager: .default
+                )
+            },
+            isExecutableFile: { FileManager.default.isExecutableFile(atPath: $0) }
+        )
+    }
+}
+
 // MARK: Construction
 
 extension TerminalSurface {
